Better support for custom SSL port
k3d uses TCP loadbalancer that can not handle HTTP headers.
So the ports on load balancer must exacly match the ports in Ingress
Controller (both on service and container) to make the X-Forwarded-Port
contain a proper value (the port used on k3d load balancer).

GoodData.CN relied on this header to compute the callback URL for
OAuth 2.0 flow.